Split TableSegment into table_segment.py
