+++ tests/common.py
@@ -40,6 +40,7 @@ if os.environ.get("LOG_LEVEL", False):
 
 logging.basicConfig(level=level)
 logging.getLogger("diff_tables").setLevel(level)
+logging.getLogger("table_segment").setLevel(level)
 logging.getLogger("database").setLevel(level)
 
 try:
